refactor: 스케줄 페이지 상태로직 분리 - 다중선택 상태와 개별상태 훅 분리 - 페이지 유틸함수 분리
